Refactored navigation from Jetpack Compose navigation to Navigation Component
