--- app/src/main/java/com/example/trashminder/presentation/createdReminder/BottomBarScreen.kt
@@ -1,31 +1,0 @@
-package com.example.trashminder.presentation.createdReminder
-
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.CalendarToday
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.ui.graphics.vector.ImageVector
-
-sealed class BottomBarScreen(
-    val route: String,
-    var title: String,
-    val icon: ImageVector
-) {
-    object Home : BottomBarScreen(
-        route = "home",
-        title = "Home",
-        icon = Icons.Default.Home
-    )
-
-    object Calendar : BottomBarScreen(
-        route = "calendar",
-        title = "Calendar",
-        icon = Icons.Default.CalendarToday
-    )
-
-    object Settings : BottomBarScreen(
-        route = "settings",
-        title = "Settings",
-        icon = Icons.Default.Settings
-    )
-}
