--- app/src/main/java/com/example/trashminder/presentation/createdReminder/Calendar.kt
@@ -1,25 +1,0 @@
-package com.example.trashminder.presentation.createdReminder
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-
-@Composable
-fun CalendarScreen() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(text = "CALENDAR")
-    }
-}
-
-@Composable
-@Preview(showBackground = true)
-fun CalendarScreenPreview() {
-    CalendarScreen()
-}
