+++ app/src/main/java/com/example/trashminder/presentation/createdReminder/HomeScreenViewModel.kt
@@ -2,27 +2,30 @@ package com.example.trashminder.presentation.createdReminder
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import com.example.trashminder.model.ListOfReminders
-import com.example.trashminder.repository.ReminderRepository
-import com.example.trashminder.utils.Response
-import kotlinx.coroutines.launch
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.firestore.FirebaseFirestore
 
-class HomeScreenViewModel: ViewModel() {
-    private val repository: ReminderRepository = ReminderRepository()
-    var reminderResponse = mutableStateOf<Response<ListOfReminders>>(Response.Loading)
+class HomeScreenViewModel : ViewModel() {
+    val reminderResponse = mutableStateOf<ListOfReminders?>(null)
+    private val firestore = FirebaseFirestore.getInstance()
+    private val auth = FirebaseAuth.getInstance()
 
     init {
         getUserData()
     }
 
     private fun getUserData() {
-        viewModelScope.launch {
-            repository.getUserData().collect { response ->
-                    reminderResponse.value = response
+        firestore.collection(collectionPath).document("${auth.currentUser?.uid}")
+            .addSnapshotListener { snapshot, e ->
+                if ((snapshot != null) && snapshot.exists()) {
+                    val data = snapshot.toObject(ListOfReminders::class.java)
+                    reminderResponse.value = data
                 }
             }
-        }
     }
 
-
+    companion object {
+        private const val collectionPath = "users"
+    }
+}
