--- app/src/main/java/com/example/trashminder/presentation/navigation/NavigationGraphs.kt
@@ -1,8 +1,0 @@
-package com.example.trashminder.presentation.navigation
-
-object NavigationGraphs {
-        const val ROOT = "root_graph"
-        const val AUTHENTICATION = "auth_graph"
-        const val MAINSCREEN = "main_screen_graph"
-        const val NEWREMINDER = "new_reminder_graph"
-}
