--- app/src/main/java/com/example/trashminder/utils/Response.kt
@@ -1,12 +1,0 @@
-package com.example.trashminder.utils
-
-sealed class Response<out T> {
-    object Loading : Response<Nothing>()
-    data class Success<out T>(
-        val data: T
-    ) : Response<T>()
-
-    data class Failure(
-        val e: Exception?
-    ) : Response<Nothing>()
-}
