+++ app/src/main/res/drawable/gradient.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+<gradient
+    android:type="linear"
+    android:startColor="#25B567"
+    android:endColor="#1C7A8E"
+    android:angle="30"/>
+</shape>
