+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+<item
+        android:id="@+id/homeFragment"
+        android:title="@string/menu_home_fragment"
+        android:icon="@drawable/ic_home"/>
+    <item
+        android:id="@+id/calendarFragment"
+        android:title="@string/menu_calendar_fragment"
+        android:icon="@drawable/ic_calendar" />
+    <item
+        android:id="@+id/settingsFragment"
+        android:title="@string/menu_settings_fragment"
+        android:icon="@drawable/ic_settings" />
+</menu>
