+++ app/src/main/res/navigation/navigation.xml
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/navigation"
+    app:startDestination="@id/loginFragment">
+
+    <fragment
+        android:id="@+id/signUpFragment"
+        android:name="com.example.trashminder.presentation.auth.signup.SignUpFragment"
+        android:label="SignUpFragment">
+        <action
+            android:id="@+id/action_signUpFragment_to_newReminderFragment"
+            app:destination="@id/newReminderFragment"
+            app:popUpTo="@+id/loginFragment"
+            app:popUpToInclusive="true"/>
+    </fragment>
+
+    <fragment
+        android:id="@+id/loginFragment"
+        android:name="com.example.trashminder.presentation.auth.login.LoginFragment"
+        android:label="LoginFragment">
+        <action
+            android:id="@+id/action_loginFragment_to_homeFragment"
+            app:destination="@id/homeFragment"
+            app:popUpTo="@id/loginFragment"
+            app:popUpToInclusive="true"/>
+    </fragment>
+
+    <fragment
+        android:id="@+id/newReminderFragment"
+        android:name="com.example.trashminder.presentation.newReminder.NewReminderFragment"
+        android:label="NewReminderFragment">
+        <action
+            android:id="@+id/action_newReminderFragment_to_homeFragment"
+            app:destination="@id/homeFragment"
+            app:popUpTo="@id/newReminderFragment"
+            app:popUpToInclusive="true"/>
+    </fragment>
+    <fragment
+        android:id="@+id/calendarFragment"
+        android:name="com.example.trashminder.presentation.createdReminder.CalendarFragment"
+        android:label="CalendarFragment"/>
+    <fragment
+        android:id="@+id/settingsFragment"
+        android:name="com.example.trashminder.presentation.createdReminder.SettingsFragment"
+        android:label="SettingsFragment">
+        <action
+            android:id="@+id/action_settingsFragment_to_loginFragment"
+            app:destination="@id/loginFragment"
+            app:popUpTo="@id/homeFragment"
+            app:popUpToInclusive="true"/>
+    </fragment>
+    <fragment
+        android:id="@+id/homeFragment"
+        android:name="com.example.trashminder.presentation.createdReminder.HomeFragment"
+        android:label="HomeFragment" >
+        <action
+            android:id="@+id/action_homeFragment_to_newReminderFragment"
+            app:destination="@id/newReminderFragment" />
+    </fragment>
+</navigation>
