+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,8 @@
 <resources>
     <string name="app_name">Trashminder</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="menu_home_fragment">Home</string>
+    <string name="menu_calendar_fragment">Calendar</string>
+    <string name="menu_settings_fragment">Settings</string>
 </resources>
